speed_test_app/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autopep8==1.6.0
Babel==2.15.0
bcc==0.18.0
bcrypt==4.1.3
beautifulsoup4==4.10.0
black==24.4.2
blinker==1.8.2
Brlapi==0.8.3
cachetools==5.5.0
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.1.7
cmake==3.30.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distutils-extra-python==2.39.2
Django==5.1.5
django-debug-toolbar==5.0.1
dnspython==2.6.1
dominate==2.9.1
email_validator==2.2.0
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
flask-babel==4.0.0
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
flask-bootstrap5==0.1.dev1
Flask-Cors==3.0.8
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Reuploaded==1.4.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
frozenlist==1.4.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.5
grpcio-status==1.62.3
gyp==0.1
h11==0.14.0
hidpidaemon==18.4.6
hiq-python==1.1.13
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
iniconfig==2.0.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.5.0
kazam==1.5.8
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.3.5
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
meson==0.61.2
more-itertools==8.10.0
multidict==6.0.5
mypy==0.942
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
mysqlclient==2.2.4
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
openai==1.42.0
outcome==1.3.0.post0
packaging==24.1
parameterized==0.9.0
pathspec==0.12.1
pep8==1.7.1
pipenv==2024.4.0
platformdirs==4.2.2
pluggy==1.5.0
pop-transition==1.1.2
proto-plus==1.24.0
protobuf==4.25.4
psutil==5.9.0
py-itree==0.0.19
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycodestyle==2.6.0
pycups==2.0.1
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic_core==2.20.1
pydbus==0.6.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
pytest==8.3.2
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.2
readmeai==0.5.82
regex==2024.7.24
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
rich==13.7.1
rsa==4.9
SecretStorage==3.3.1
selenium==4.27.1
sessioninstaller==0.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.3.1
speaklater==1.3
speedtest-cli==2.1.3
SQLAlchemy==2.0.31
sqlparse==0.4.2
systemd-python==234
tenacity==8.5.0
tiktoken==0.4.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.5
trio==0.28.0
trio-websocket==0.11.1
typed-ast==1.4.3
types-aiofiles==0.1
types-annoy==1.17
types-appdirs==1.4
types-atomicwrites==1.4
types-aws-xray-sdk==2.8
types-babel==2.9
types-backports-abc==0.5
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.10
types-bleach==4.1
types-boto==2.49
types-braintree==4.11
types-cachetools==4.2
types-caldav==0.8
types-certifi==2020.4
types-characteristic==14.3
types-chardet==4.0
types-click==7.1
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-contextvars==0.1
types-croniter==1.0
types-cryptography==3.3
types-dataclasses==0.1
types-dateparser==1.0
types-DateTimeRange==0.1
types-decorator==0.1
types-Deprecated==1.2
types-docopt==0.6
types-docutils==0.17
types-editdistance==0.5
types-emoji==1.2
types-entrypoints==0.3
types-enum34==1.1
types-filelock==3.2
types-first==2.0
types-Flask==1.1
types-freezegun==1.1
types-frozendict==0.1
types-futures==3.3
types-html5lib==1.1
types-httplib2==0.19
types-humanfriendly==9.2
types-ipaddress==1.0
types-itsdangerous==1.1
types-JACK-Client==0.1
types-Jinja2==2.11
types-jmespath==0.10
types-jsonschema==3.2
types-Markdown==3.3
types-MarkupSafe==1.1
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.0
types-oauthlib==3.1
types-orjson==3.6
types-paramiko==2.7
types-Pillow==8.3
types-polib==1.1
types-prettytable==2.1
types-protobuf==3.17
types-psutil==5.8
types-psycopg2==2.9
types-pyaudio==0.2
types-pycurl==0.1
types-pyfarmhash==0.2
types-Pygments==2.9
types-PyMySQL==1.0
types-pyOpenSSL==20.0
types-pyRFC3339==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-dateutil==2.8
types-python-gflags==3.1
types-python-nmap==0.6
types-python-slugify==5.0
types-pytz==2021.1
types-pyvmomi==7.0
types-PyYAML==5.4
types-redis==3.5
types-requests==2.25
types-retry==0.9
types-selenium==3.141
types-Send2Trash==1.8
types-setuptools==57.4
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-stripe==2.59
types-tabulate==0.8
types-termcolor==1.1
types-toml==0.10
types-toposort==1.6
types-ttkthemes==3.2
types-typed-ast==1.4
types-tzlocal==0.1
types-ujson==0.1
types-vobject==0.9
types-waitress==0.1
types-Werkzeug==1.0
types-xxhash==2.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.28.1
visitor==0.1.3
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
